requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
decorator==5.2.1
distro==1.9.0
et_xmlfile==2.0.0
executing==2.2.0
Flask==3.1.1
frozenlist==1.6.0
genai==2.1.0
google-auth==2.40.2
google-genai==1.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==8.36.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multidict==6.4.4
openai==0.27.10
openpyxl==3.1.5
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.1
PyMuPDF==1.26.0
pytesseract==0.3.13
python-docx==1.1.2
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
stack-data==0.6.3
tabulate==0.9.0
tiktoken==0.3.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.0
